Fix build revision in build output
We still used the svn-revision-number-maven-plugin to figure out a build
ID. Since we've switched to git in the mean time, that no longer makes
sense.

Use the GitDescribe-Maven-Plugin instead. That plugin as written insist
on using a prefix that we don't want, so we've (currently) patched it
and are using the patched version. Once the pull request has been
accepted and a new version released, we'll update the version -- but for
now, use a -SNAPSHOT version.