Require JVM 1.8+
There are issues with JVM 1.7, so allowing the script to select that
seems like a bad idea.

Unfortunately, we can't declare that Java 6 is allowed if we say that
Java 7 isn't. On the other hand, Java 6 has not been supported for quite
a while now, so not sure if it's that much of a loss anymore...