Adjust JRE version requirements
There is no point in bumping the maximum version to a version that
doesn't exist yet. If there ever is a version that we know this doesn't
work with, we can fix it then. Until then, just drop the "maximum
supported version" from the pom.xml.

Also, bump the minimally required version from 1.6 to 1.8, so we're in
sync with OS X.