Fix certificate fetching
When we try to read a certificate, the code would ask the PC/SC layer to
pull it either from its cache or from the card. Unfortunately, this
failed to account for the possibility that the data was loaded from file
rather than from a card.

Change the code so that iff the data was loaded from file, we read
certificates from the certificate chains, rather than trying to read
them from nonexisting cards.